feat: Moves to components folder
