+++ src/components/DailyTemperature/DailyTemperature.stories.js
@@ -1,0 +1,18 @@
+import React from 'react'
+
+import DailyTemperature from '@components/DailyTemperature/DailyTemperature'
+import { View } from 'react-native'
+
+export default {
+  title: 'DailyTemperature',
+  component: DailyTemperature,
+  decorators: [
+    (Story) => (
+      <View style={{ backgroundColor: 'red' }}>
+        <Story />
+      </View>
+    )
+  ]
+}
+
+export const Basic = {}
