+++ src/components/Loading/Loading.stories.js
@@ -1,0 +1,18 @@
+import React from 'react'
+
+import { LoadingRequest } from '@components/Loading/LoadingRequest'
+import { View } from 'react-native'
+
+export default {
+  title: 'Loading',
+  component: LoadingRequest,
+  decorators: [
+    (Story) => (
+      <View >
+        <Story />
+      </View>
+    )
+  ]
+}
+
+export const Basic = {}
